requirements.txt
asgiref==3.4.0
Django==3.2.4
django-crispy-forms==1.12.0
django-filter==2.4.0
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
yapf==0.31.0
